+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - [feat:support ipv6.](https://github.com/Tencent/spring-cloud-tencent/pull/1601)
 - [feat:support config all recover enabled.](https://github.com/Tencent/spring-cloud-tencent/pull/1605)
 - [feat:support stat reporting path aggregation.](https://github.com/Tencent/spring-cloud-tencent/pull/1609)
+- [feat:support namespace exports configuration if not created.](https://github.com/Tencent/spring-cloud-tencent/pull/1614)